+++ featherstore/_table/common.py
@@ -14,7 +14,8 @@ def can_init_table(table_name, store_name):
     current_db()
 
     if not isinstance(table_name, str):
-        raise TypeError(f"'table_name' must be a str, is type {type(table_name)}")
+        raise TypeError(
+            f"'table_name' must be a str, is type {type(table_name)}")
 
     if table_name == METADATA_FOLDER_NAME:
         raise ValueError(f"Table name {METADATA_FOLDER_NAME} is forbidden")
@@ -29,7 +30,8 @@ def can_init_table(table_name, store_name):
 
 def can_rename_table(new_table_name, old_table_path, new_table_path):
     if not isinstance(new_table_name, str):
-        raise TypeError(f"'new_name' must be a str (is type {type(new_table_name)})")
+        raise TypeError(
+            f"'new_name' must be a str (is type {type(new_table_name)})")
 
     if new_table_name == METADATA_FOLDER_NAME:
         raise ValueError(f"Table name {METADATA_FOLDER_NAME} is forbidden")
@@ -53,7 +55,7 @@ def combine_partitions(partitions):
 def format_table(df, index, warnings):
     df = _convert_to_pandas(df)
     df = _set_index(df, index)
-    _check_index_constraints(df)
+    _check_index_constraints(df.index)
 
     index_is_sorted = df.index.is_monotonic_increasing
     if not index_is_sorted:
@@ -92,14 +94,6 @@ def _set_index(df, index_name):
     if df.index.name == DEFAULT_ARROW_INDEX_NAME:
         df.index.name = None
     return df
-
-
-def _check_index_constraints(df):
-    index_type = df.index.inferred_type
-    if index_type not in {"integer", "datetime64", "string"}:
-        raise TypeError("Table.index type must be either int, str or datetime")
-    if df.index.has_duplicates:
-        raise ValueError("Values in Table.index must be unique")
 
 
 def _sort_index(df, warnings):
@@ -165,3 +159,19 @@ def _get_cols(df, has_default_index):
         if has_default_index and DEFAULT_ARROW_INDEX_NAME not in cols:
             cols.append(DEFAULT_ARROW_INDEX_NAME)
     return cols
+
+
+def _check_index_constraints(index):
+    if index.has_duplicates:
+        raise IndexError("Values in Table.index must be unique")
+    index_type = index.inferred_type
+    if index_type not in {"integer", "datetime64", "string"}:
+        raise TypeError("Table.index type must be either int, str or datetime")
+
+
+def _check_column_constraints(cols):
+    cols = pd.Index(cols)
+    if cols.has_duplicates:
+        raise IndexError("Column names must be unique")
+    if "like" in cols.str.lower():
+        raise IndexError("df contains invalid column name 'like'")